requirements.txt
Flask==3.0.3
mysql-connector-python==8.4.0
bcrypt==4.1.3
